implement matched order publisher

+++ include/matching/L2Data.hpp
@@ -3,8 +3,8 @@
 #include "matching/Order.hpp"
 struct alignas(64) L2Data {
     uint8_t instrument_id_;
-    uint8_t price_level_;
-    int32_t quantity_;
+    uint64_t price_level_;
+    int64_t quantity_;
     OrderSide side_;
 
     bool operator==(const L2Data&) const = default;
